tmp_rovodev_rag_app/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
requests==2.31.0
numpy>=1.24.0
sentence-transformers>=2.7.0
PyPDF2==3.0.1
python-docx==1.1.0
tiktoken>=0.5.0
openai>=1.0.0
streamlit==1.28.1
huggingface-hub>=0.19.0
transformers>=4.36.0
torch>=2.0.0
tokenizers>=0.15.0
msgpack>=1.0.5
python-dotenv>=0.19.0